Add support for downloading important files #8
* Modified comments and packages, classes and functions names
* Removed parameter from downloadArchiveModal
