+++ application-admintools-default/src/main/java/com/xwiki/admintools/internal/files/resources/DataProvidersDataResource.java
@@ -17,7 +17,7 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package com.xwiki.admintools.internal.download.resources;
+package com.xwiki.admintools.internal.files.resources;
 
 import java.io.IOException;
 import java.util.HashMap;
