+++ application-admintools-ui/src/main/resources/AdminTools/Code/AdminToolsJS.xml
@@ -124,9 +124,9 @@
     </property>
     <property>
       <code>require(['jquery'], function($) {
-  $(document).on('click', '#filesDownloadFilesModal .btn-primary', function(event) {
+  $(document).on('click', '#DownloadFilesModal .btn-primary', function(event) {
     event.preventDefault();
-    const downloadForm = $('#filesDownloadFilesModal form');
+    const downloadForm = $('#DownloadFilesModal form');
 
     // Simulate a click on a link to initialize the download.
     const link = document.createElement('a');
@@ -142,7 +142,7 @@
     let noLines = modal.find("input[name='noLines']").val();
 
     // If the input is empty, the default value of 1000 will be requested.
-    if(noLines == '') {
+    if (noLines == '') {
       noLines = '1000';
     }
 
@@ -153,11 +153,11 @@
     link.click();
   });
 
-  $(document).on('change', '#filesDownloadFilesModal input[name="files"][value="logsDataResource"]', function() {
+  $(document).on('change', '#DownloadFilesModal input[name="files"][value="logsDataResource"]', function() {
     if ($(this).is(':checked')) {
-        $('#filesDownloadFilesModal .dateFields').show();
+        $('#DownloadFilesModal .dateFields').show();
     } else {
-        $('#filesDownloadFilesModal .dateFields').hide();
+        $('#DownloadFilesModal .dateFields').hide();
     }
   });
 });</code>
